+++ test.py
@@ -1,0 +1,4 @@
+''' Run required tests'''
+from tests.library import test_scan
+
+test_scan(3)